Refactor token expiry handling in authentication responses
- Updated the calculation of `expires_in` and `refresh_threshold` in the login, registration, email verification, password reset, and SSO completion endpoints to use environment variables for consistency.
- Ensured that the token expiry and refresh threshold values are dynamically set based on configuration, improving maintainability and flexibility.

+++ robosystems/routers/auth/login.py
@@ -210,8 +210,8 @@ async def login(
   # Calculate token expiry and refresh threshold
   from ...config import env
 
-  expires_in = int(env.JWT_EXPIRY_HOURS * 3600)  # Convert hours to seconds
-  refresh_threshold = env.TOKEN_GRACE_PERIOD_MINUTES * 60  # Convert minutes to seconds
+  expires_in = int(env.JWT_EXPIRY_HOURS * 3600)
+  refresh_threshold = int(env.TOKEN_GRACE_PERIOD_MINUTES * 60)
 
   return AuthResponse(
     user={
